requirements.txt
pefile==2023.2.7
py4j==0.10.9.7
pyspark==3.5.0
python-dotenv
findspark
psycopg2-binary
python-logstash-async
elasticsearch
boto3==1.28.57
botocore==1.31.57
pytest
pytest-mock


